support namespaced user info for streams
Makes use of a new stream metadata field "labelledRefs" which stores
sibling object references to a stream object as "labels" consisting
of key-value pairs with unique keys. Data about the user "queueing"
a stream is stored as a labelledRef, with the user's room's UUID as
the unique label key, and the user object as the value.